refactor: host shared timing types and RuleScope in common
common/artifact.ts imported its timing types from the report layer, and
common/diagnostic.ts pulled RuleScope from the engine, anchoring the two
most-shared type modules to heavier layers. The types now live in
common/; report/timings.ts and engine/rules/types.ts re-export them so
existing imports hold.
