+++ packages/nestjs-doctor/src/common/rule-scope.ts
@@ -1,0 +1,1 @@
+export type RuleScope = "file" | "project" | "schema";
